+++ debounce-form-sample/src/components/SearchForm.tsx
@@ -1,22 +1,13 @@
-import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
-
-interface Props {
-  setDebouncedValue: Dispatch<SetStateAction<string>>;
-}
+import React, { useState } from "react";
+import { useDebounce } from "../hooks/useDebounce";
+import { useFetchQiitaItems } from "../hooks/useFetchQiitaItems";
+import { SearchResult } from "./SearchResult";
 
 // 検索フォームコンポーネント（ここで debouncedValue を更新する）
-export const SearchForm: React.VFC<Props> = (props) => {
+export const SearchForm: React.VFC = () => {
   const [formText, setFormText] = useState<string>("");
-
-  // formText の変更を検知して実行される Effect
-  useEffect(() => {
-    // 表示までのtimerをセット、１秒間フォームの入力動作が行われなければ
-    // フォームの値で debouncedValue を更新する
-    const timer = setTimeout(() => props.setDebouncedValue(formText), 1000);
-
-    // 次の Effect が実行される直前に timer をキャンセル
-    return () => clearTimeout(timer);
-  }, [formText]);
+  const debouncedValue = useDebounce(formText, 1000);
+  const qiitaItems = useFetchQiitaItems(debouncedValue);
 
   return (
     <>
@@ -29,6 +20,7 @@ export const SearchForm: React.VFC<Props> = (props) => {
         value={formText}
         onChange={(e) => setFormText(e.target.value)}
       />
+      <SearchResult qiitaItems={qiitaItems} />
     </>
   );
 };
